requirements.txt
Flask
pymongo
flasgger
certifi
flask-bcrypt
python-dotenv
PyJWT